refactoring: 리뷰 수정 MVVM으로 변경
